Move change set tests to v2
